requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
pandas==1.4.3
pymusixmatch==0.3

